refactor: update Winner model and related mappers for improved data handling and add test cases for bulk insert functionality

+++ chap05/src/main/java/com/genius/primavera/application/WinnerService.java
@@ -34,7 +34,7 @@ public interface WinnerService {
 
 	List<Winner> findAllCommitted();
 
-	Winner findAllByIdReadCommitted(int id);
+	Winner findAllByIdReadCommitted(Long id);
 
-	Winner findAllByIdRepeatableRead(int id);
+	Winner findAllByIdRepeatableRead(Long id);
 }
